requirements.txt
pdfminer.six
pypandoc
langchain-community 
langchain-openai 
langchain-chroma
pydantic
chromadb 
python-dotenv
tiktoken
